requirements.txt
django<6.1
djangorestframework
django-cors-headers
mysqlclient
python-dotenv
django-unfold